DEV-95: Implemented the special scroll in a custom hook
